Prepare some of X86Target for 32-bit PC

+++ core/compiler/X86Target.cpp
@@ -361,7 +361,7 @@ void X86Target::init(SourceInfo sourceInfo, void *cpuPtr)
     this->cpuPtr = cpuPtr;
 }
 
-bool X86Target::compile(uint8_t *&codePtr, uint8_t *codeBufEnd, uint16_t pc, GenBlockInfo &blockInfo)
+bool X86Target::compile(uint8_t *&codePtr, uint8_t *codeBufEnd, uint32_t pc, GenBlockInfo &blockInfo)
 {
     X86Builder builder(codePtr, codeBufEnd);
 
@@ -369,8 +369,8 @@ bool X86Target::compile(uint8_t *&codePtr, uint8_t *codeBufEnd, uint16_t pc, Gen
 
     // state
     uint8_t *lastInstrCycleCheck = nullptr;
-    std::map<uint16_t, uint8_t *> branchTargets;
-    std::multimap<uint16_t, uint8_t *> forwardBranchesToPatch;
+    std::map<uint32_t, uint8_t *> branchTargets;
+    std::multimap<uint32_t, uint8_t *> forwardBranchesToPatch;
     int needCallRestore = 0;
